report: extract the deterministic core into the package (#251, slice 1)
Codex's bar for this work: a CLI that shells out to
scripts/harness/report_postflight.py is a rename, not independence. So this
moves the phase logic into importable API rather than wrapping it.

- clawock/validation.py — the text/numeric validation primitives (markdown
  tables, forbidden phrases, numeric-claim gate, advisory/critical
  categorisation, report char limits and prose budget), moved out of
  scripts/harness/_harness_common.py.
- clawock/report.py — assemble_message, validate, categorize and the section /
  forbidden-phrase / char-limit constants, moved out of report_postflight.py.
- Both old modules re-export what moved, so all ten in-repo importers and the
  twenty-odd report regression tests are untouched and still exercise exactly
  this code.
- `clawock report --context CTX --prose FILE [--json]` runs that core in-process.

Delivery and publication are deliberately NOT part of this command. They are
capability providers; a report you can render without them is the point.

Proof it is not a facade: the report renders from an installed wheel, in a
temp cwd outside any checkout, with PATH holding no openclaw, git or gh and
scripts/ not importable. Verified both as a test and by hand against a real
built wheel.

Two tests, mutation-verified: the no-openclaw/no-git/no-scripts run, and an AST
check that the package core never imports subprocess (adding one reds it — the
modules legitimately mention report_postflight.py in prose, so text matching
would not do).

Refs #251
